# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ultra lightweight configuration data service",
        "title": "Source",
        "contact": {
            "name": "SouthWinds Tech Ltd",
            "url": "https://southwinds.io/",
            "email": "info@southwinds.io"
        },
        "version": "1.0"
    },
    "paths": {
        "/item": {
            "get": {
                "description": "Get all the configurations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get all the configurations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item/pop/newest/{type}": {
            "delete": {
                "description": "Get the newest configuration that have the specified type and remove it from the database effectively acting as a LIFO queue",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get the newest configuration that have the specified type and remove it from the database effectively acting as a LIFO queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the type of the configuration to pop",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item/pop/oldest/{type}": {
            "delete": {
                "description": "Get the oldest configuration that have the specified type and remove it from the database effectively acting as a FIFO queue",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get the oldest configuration that have the specified type and remove it from the database effectively acting as a FIFO queue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the type of the configuration to pop",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item/tag/{tags}": {
            "get": {
                "description": "Get all the configurations that have the specified tags",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get all the configurations that have the specified tags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "a pipe separated list of tags (e.g. tag1|tag2|tag3) where tag is the tag name, not the value",
                        "name": "tags",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item/type/{type}": {
            "get": {
                "description": "Get all the configurations that have the specified type",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get all the configurations that have the specified type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the type of the configurations to retrieve",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item/{key}": {
            "get": {
                "description": "Get value of a configuration item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get the value of a configuration item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the key for the configuration item to get",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Set value of a configuration item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Set the value of a configuration item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the key for the configuration item to set",
                        "name": "key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the json based configuration",
                        "name": "schema",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "the key that defines the type of item for validation purposes. If not specified, no validation is performed.",
                        "name": "Source-Type",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a configuration item",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Delete a configuration item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the key for the configuration item to delete",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/item/{key}/children": {
            "get": {
                "description": "Get the children linked to a configuration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get the children linked to a configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the key for the item having the children",
                        "name": "key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}